fix commit 4.1

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -58,7 +58,7 @@ public class StudentController {
     }
 
     @GetMapping("avg")
-    public Integer findAvgAgeStudent() {
+    public Float findAvgAgeStudent() {
         return studentService.findAvgAgeStudent();
     }
     @GetMapping("5student")
